[SH] Crash fix: Nulls checks before getting the anchor
In https://chromium-review.googlesource.com/c/chromium/src/+/3230201
(landed in 97.0.4690.0) we added a code that will end up requesting
TextFragmentAnchor for all URL that have text fragment. It is possible
that TextFragmentAnchor is requested for frames or views that already
got deattached, hence the crash.
It is also likely that we will have similar situation in other code
paths that request TextFragmentAnchor but due to them being much more
rare we never caught this case.

Bug: 1266976
Change-Id: Ib844529a28548ce69763da16fbb2b452a599ccff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3282805
Reviewed-by: sebsg <sebsg@chromium.org>
Reviewed-by: Cheick Cisse <cheickcisse@google.com>
Commit-Queue: Gayane Petrosyan <gayane@google.com>
Cr-Commit-Position: refs/heads/main@{#941825}
NOKEYCHECK=True
GitOrigin-RevId: 98e9eda2168f7c016aab467528d1093f8f0153a3