[LR] Fix BaseGifDrawable reset race condition
Attempts to fix threading and timing issues around resetting the GIF
decoding operations of Lightweight Reactions. These timing issues are
believed to be the cause of crashes that could happen when duplicating
a reaction, adding a new reaction to the scene, changing a reaction's
type, and starting the GIF generation.

The specific changes are:

- Keep track of whether a decode operation is currently in progress in
  the ReactionGifDrawable subclass.
- When stepping mode is first enabled in ReactionGifDrawable, wait for
  any in-progress decode operation to finish before starting the steps,
  to ensure the next frame we get is truly part of the step() and not
  part of the regular animation that was playing before stepping mode
  was enabled.
- Do not call BaseGifDrawable.reset() on the UI thread. Instead, add a
  new public method requestReset() in BaseGifDrawable that posts the
  reset request to the decoder thread, ensuring the data read position
  is not moved while the decoder thread is reading the next value.

There is no way to really confirm the crashes were caused by these
thread timing issues, since the crashes are intermittent and rare, but
based on a few stack traces I've managed to capture, I have good
confidence at least some of the crashes will be fixed by this CL. This
CL also seems to have fixed a small visual glitch in the generated GIF
where the first frame of some reactions seemed out of place. That was
probably caused by an in-progress decode operation as stepping mode
was enabled, which means the first frame we fed to the encoder was
the frame of the regular animation that had been playing rather than
the frame from the controlled steps.

Bug: 1268172
Change-Id: Ic87de5863498bd32713cad8eac389b07303c03b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3285266
Reviewed-by: Sophey <sophey@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/main@{#942304}
NOKEYCHECK=True
GitOrigin-RevId: 684f2c7a9878d3b08250bc13c6803c5f69fd2b12