ScrollbarThemeMac: Hard-code NSScrollerImp constants
There are several constants related to the geometry of the scrollbar
that are queried from the private NSScrollerImp interface. These
constants depend on the type (overlay vs not) and size (regular vs
small) of the scrollbar.

Towards removing all Cocoa code, hard-code the values for each
combination, instead of determining them dynamically. Of note is that
several of these values are already duplicated (with slightly different
values) in NativeThemeMac, which is what ultimately does the drawing.

The motivation to remove all Cocoa code is to reduce our use of
undocumented APIs, and to allow the renderer process sandbox to be
tightened.

Bug: 961835
Change-Id: I80a91e7e4e1927de294f8151a90028b7a34e5e12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3271499
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942337}
NOKEYCHECK=True
GitOrigin-RevId: 7a654d8fec3cb06b4bba5e87711c0b7eff7a1026