Move PowerStatusHelper creation to WebMediaPlayerImpl
MediaFactory directly instantiates PowerStatusHelper. This requires that
class to be part of the public Blink API. In order to reduce that API,
the creation of PowerStatusHelper is moved to WebMediaPlayerImpl,
removing its usage in content/.

This also fixes an usage of a Chromium Mojo interface, replacing it with
the Blink variant, now that PowerStatusHelper is Blink-only.

Bug: 1198318
Change-Id: I951601395c334891bb6b7f486615c9486d3573cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2923208
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Will Cassella <cassew@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/main@{#942434}
NOKEYCHECK=True
GitOrigin-RevId: c3fe538055c0065f5ddedb9dac7b3bc425794f40