[mq-4] Add MediaQueryExpNode::Not/Nested/And/Or utils
These utils simply return a MediaQueryExpNode of the relevant type
while handling nullptr gracefully. This makes it possible to parse
<media-and/or> will less code.

This is worthwhile because the grammar for container queries also use
the same not/and/or pattern [1], and these utils makes that
boilerplate seem more acceptable.

[1] https://drafts.csswg.org/css-contain-3/#typedef-container-condition

Bug: 962417
Change-Id: I111e54980b27314b380e19e5bd67312b853b777b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3282139
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942522}
NOKEYCHECK=True
GitOrigin-RevId: 2bb0a0e02f597a26bc651227d6e8681cd9f74f66