[mq-4] Support range syntax in media queries
Range syntax in Media Queries Level 4 has multiple forms:

 1. <mf-name> on the left side, e.g: width > 10px
 2. <mf-name> on the right side, e.g: 10px < width, or
 3. As a bounded interval, e.g.: 5px < width < 10px.

This produces quite complicated parsing code, since the <mf-name>
can appear in multiple places, and we need to know the <mf-name>
in question *before* we can parse a value for it. This is
a problem for (2) and (3), since the <mf-name> appears *after* the
first value.

This CL implements the parsing by splitting the tokens on the
top-level separators (:,<,<=,=,=>,>) into several segments. We can
then try to match segments individually against <mf-name>/<mf-value>.
This relies on the fact that no <mf-value> contains top-level
separator tokens. If such values appear in the  future, the approach
must be reconsidered.

Bug: 962417
Fixed: 1034465
Change-Id: Ieba8e560f30f92134b7d7f8070eb5067b5edaf9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3269897
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#942578}
NOKEYCHECK=True
GitOrigin-RevId: a57094077b2b5504959e37bb192a6ac935c39920