Log both TabSwitchResult and TabSwitchResult2 during experiment
It's awkward to compare different histograms between experiment groups.
During the TabSwitchMetrics2 experiment, log both TabSwitchResult2 and
TabSwitchResult with the same values, so that it's easy to compare the
same TabSwitchResult metrics for the control group and experiment group.
If TabSwitchResult in the experiment group is ok, so is TabSwitchResult2,
and we can then deprecate TabSwitchResult since the historical data
outside the experiment group is unreliable.

R=dgozman

Bug: 1164477
Change-Id: Ie3e1857be29167e8147f883e8d628e759021c959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3286769
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#942617}
NOKEYCHECK=True
GitOrigin-RevId: e21bf9cf9168aedc46df5ed3671c2c17b6af8fe6