[rust] Rearrange stdlib source code layout
gn can inform IDEs such VSCode to find Rust code. That includes
the Rust stdlib, which doesn't currently get found by IDEs in Chromium.

This is the first in a series of two CLs which make that work.
In this CL, we rearrange the layout of stdlib source code (taken from
the Android Rust toolchain) to correspond to that of Fuchsia's Rust
toolchain, which is what gn assumes.

A subsequent CL will absorb the revised CIPD package and adjust
gn rules too.

Bug: 1269992
Change-Id: Ia3d98c07807fc106e58197ea8a1172cb2209b004
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3287261
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942650}
NOKEYCHECK=True
GitOrigin-RevId: e363ec886d193f975145ad371f7156955f4a9320