[inert] Recalc fullscreen element style when opening modal dialog
When there is a modal dialog or a fullscreen element, all the elements
are considered to be inert, except for the dialog or fullscreen element
and its subtree.

But it's possible to simultaneously have both a modal dialog and a
fullscreen element. In that case, the fullscreen element also becomes
inert, and the dialog doesn't.

However, opening or closing a modal dialog wasn't triggering a style
recalc to the fullscreen element, so dynamic cases could have a wrong
inertness.

Note that's not necessary if we already had a modal dialog and then we
open a new one. In that case, the fullscreen element would have already
been inert.

Also, if we open a modal dialog with no prior one, then the inertness of
the dialog won't change, so we don't have to recalc its style. Same when
making an element fullscreen, if there was no other fullscreen element.

Bug: 692360

TEST=StyleResolverTest.IsInertWithDialogAndFullscreen

Change-Id: I20621a219801077173a5524495556a7fd990a810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3282272
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/main@{#942700}
NOKEYCHECK=True
GitOrigin-RevId: f4b6354a57a403a0b56b940104013cda4c37a7af