Fix excessive caching of ::-webkit-scrollbar-* style in ComputedStyle
crrev.com/c/2026694 caches custom scrollbar styles on the scroll
container's ComputedStyle, so that we can conveniently check whether
any scrollbar style depends on font metrics and hence needs invaliation
when fonts change.

However, it introduced a regression that whenever scrollbar part style
changes, it adds a copy of the style to the cache. And if the scroll
container is never invalidated, the cache will keep growing, which is
not correct.

This patch fixes it by stopping caching scrollbar styles, and using a
flag in ElementRareData to track if any scrollbar style depends on
font metrics.

Fixed: 1270190
Change-Id: Id7bbe9d06ef22ffc48845555bf3d7f3bcaa0ef16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3283843
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942730}
NOKEYCHECK=True
GitOrigin-RevId: 5cef812ec8f8772a17e1a21e5e6c121e3377b539