Add Blink test to verify how cookie partitioning interacts with ancestor chain
The Partitioned attribute should only partition cookies by the site of
the topmost level frame. Meaning that if the top-level site is a.com,
then an a.com frame embedded in a 3P frame should still have access
to the same cookie partition as code running in a.com's top-level
context.

We also add a Partitioned cookie with SameSite=Lax to make sure that
it is not accessible from an a.com frame embedded in a 3P frame. Unlike
Partitioned, the SameSite attribute should consider the entire ancestor
chain when determining site for cookies. This test ensures that the
Partitioned attribute did/will not introduce a regression for the
existing semantics of the SameSite attribute.

Bug: 1225444
Change-Id: I5d107e3a241bbce90f596fc8d26a54655fe72478
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3286766
Reviewed-by: Steven Bingler <bingler@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/main@{#942731}
NOKEYCHECK=True
GitOrigin-RevId: 4e3065e9896afe69bac718ecc78e53c6083437ec