Check distributed_point_functions library in directly
The library is moved from being checked in via DEPS to being checked
into the code base directly. This enables a dependency on //base in
glog/logging.h (instead of redefining DCHECK), which allows for library
headers containing DCHECKs to be included without macro redefinition
errors. This will be required to roll the library to the latest version.
To avoid cherry picking errors with this change, the source code
directory name is changed from src/ to code/.

Bug: 1227772
Change-Id: Iae9c337485d7af0cce216336f08b1e27fbbe0567
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3217256
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942882}
NOKEYCHECK=True
GitOrigin-RevId: 9351d1b6a628ca587f50795d0cf93da3090b1793