SVG Text NG: Fix text layout after ancestor transform change
If 'transform' property of an ancestor is changed, and <text> is
painted without laying out, the content was painted at a wrong position
because the <text> kept positions computed with the old scaling factor.

This CL fixes it by invalidating <text> layout on 'transform' changes
on any ancestors.

Note: Why the legacy SVG <text> didn't need the invalidation?
Actually it's necessary for the legacy SVG <text> too. Fortunately the
legacy SVG <text> stores unscaled positions, and paint text on the
almost right positions. However the positions were computed with the
old scaled fonts, and they are not ideal results with the new scaling
factor.

Bug: 1270713
Change-Id: I2d972a16f1203bb3505c3ac4d78388167b6ee383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3289567
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942976}
NOKEYCHECK=True
GitOrigin-RevId: f90847e44cc72d9b23ebfa979aaf892afd52942f