MediaStreamAudioProcessor: Log messages via a callback
This CL makes the main logging function in MediaStreamAudioProcessor injectable, in order to make the code usable from the audio process as well as the renderer process.

Note that the processing functions (ProcessData, OnPlayoutData, ...) do not use MediaStreamAudioProcessor::SendLogMessage. This is either for some legacy reasons, or a mistake by the person (me!) who added those log lines. While it is possible to fix this right now, since SendLogMessage ends up calling WebRtcLogMessage which can be called on any thread, this might not work in the audio process. Leaving those log lines as LOG() lets us restrict audio process logging to the control thread.

In the audio process, the current plan is to forward logs to the InputController::EventHandler:
https://source.chromium.org/chromium/chromium/src/+/main:services/audio/input_controller.h;l=99;drc=65c9257f1777731d6d0669598f6fe6fe65fa61d3

Bug: chromium:1215061
Change-Id: I6c0a95a84af7158f0d57929c136a570cac6e2411
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3263929
Commit-Queue: Sam Zackrisson <saza@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Reviewed-by: Henrik Andreasson <henrika@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942999}
NOKEYCHECK=True
GitOrigin-RevId: 4529043f80cc27fd311f67e0360772c22af0d764