SHELL: Support component build for SHELL
This CL is adding component build support for SHELL library.
That change is needed in order for the library to be used by
any number of consumers. That will be needed in order for
the private_membership library to have its component build
as well.

This CL is doing the following:
1) Allowing for the library to use component macros needed to
export the classes symbol correctly
2) Adding a patch set with all these library's changes
3) Adding explicit definitions for the generic templates classes

Also, these changes will be modified later on the upstream
with correct macro that will be mapped to empty. That is needed
now in order to unblock another team to use the private_membership
library to avoid further delay for their launch.

Bug: 1241832, b/201763017

Change-Id: I651d50ae579bd72a5459a59892549ec4eb250575
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3225925
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943022}
NOKEYCHECK=True
GitOrigin-RevId: f2bcadd8fbfe76a0bc7169693393b2086863006a