Implement |NGFragmentItem::FragmentId| for line boxes
Paint uses |FragmentId| for the key of |DisplayItemClient|.
It did not support line boxes, as it was easy for the painter
to compute it, until block fragmentation is supported.

This patch computes |FragmentId| for line boxes, and change
the painter to use it.

The test:
  fast/forms/select/listbox-in-multi-column.html
started to hit DCHECK when NG block fragmentation is enabled
since CompositeAfterPaint is enabled. This patch fixes it.

Bug: 1270261
Change-Id: Ia51e4632233aafaee1988afbb0460c47553b227e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3288511
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943024}
NOKEYCHECK=True
GitOrigin-RevId: f258745f5a9ada564acab3e081442d4deed9352e