Fail Deserialization for malformed 1P StorageKeys
A first-party (1P) StorageKey should always serialize to the format
`origin`, specifically excluding the top-level site (since it's
same-site to the origin).

Because this is the case, the deserialization will now fail if it
encounters a serialized string for a 1p StorageKey that includes the
top-level site as that would indicate a malformed input.

Valid:
https://example.com
https://example.com^0https://notexample.com

Malformed:
https://example.com^0https://example.com

Bug: 1199077
Change-Id: I8f88d10cfad1a5626cb2e8609d3cdea8cb27a3e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291632
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943263}
NOKEYCHECK=True
GitOrigin-RevId: 51dba2713ece22282912c4d15006f05c7266edd4