[FlexNG] Adjust flexbox info for item expansion
When a flex-item expands due to fragmentation, we need to adjust the
intrinsic-block-size and subsequent offsets.

To accomplish this, store the original block-size of each item on
NGFlexItem, and reduce the original block-size as the item is
fragmented. If this block-size becomes negative, we know the item
expanded as a result of fragmentation.

This expansion is then accumulated within the NGFlexLine where it
can be used to adjust the offsets of subsequent items.

LayoutWithBlockFragmentation() was also removed, and its logic
was moved into GiveItemsFinalPositionAndSizeForFragmentation() for
convenience.

Bug: 660611
Change-Id: I67524cd436a87aebd52382682b2cea9d9d32b018
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3290486
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943323}
NOKEYCHECK=True
GitOrigin-RevId: b38d348456681ac077b7abea1733b5d3f080cb94