Fix various bugs with vector printing of 2D canvas content
The primary objective of this CL is to resolve problems of missing
content in printed web pages caused by the canvas auto-flushing
mechanism.  While writing unit tests to improve coverage of vector
printing use cases, several other issues were discovered which are also
fixed in this CL:
  * CompositeAfterPaint was preventing vector printing from working
    in cases where the canvas has its own compositing layer.
  * The flushing that happens when a source image is touched after a
    drawImage call, was causing missing content. (only affects OOPR)

The reason for the missing content was that Canvas2DLayerBridge was
tracking the clear_frame_ state, which tells us when vector printing is
not possible.  The problem with this is that there are several (new)
code paths that trigger flushes in the CanvasResourceProvider without
notifying the Canvas2DLayerBridge, which caused the clear_frame_ state to
be wrong.

BUG=1267814

Change-Id: Ie802c4f24e49cf16c886e01e8b643553b3179d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3288019
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943593}
NOKEYCHECK=True
GitOrigin-RevId: b38a9930001ab4a348064ca3d2d9f3f92f743f9e