Always forward VisualProperties from UpdateViewportIntersection
The call to SynchronizeVisualProperties() from RemoteFrameView::
UpdateViewportIntersection() entirely defeats the purpose of sending
VisualProperties together with ViewportIntersection, which is to
ensure that changes to both structs are applied atomically in the
remote renderer.

Removing that call exposed another bug, which is that changes to
VisualProperties that are sent along with ViewportIntersection may
be dropped for top-level frames, such as portals.

Bug: chromium:1135714
Change-Id: Ie05c6b202b30f83edddb3934cfe48c063acdac3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291610
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943714}
NOKEYCHECK=True
GitOrigin-RevId: 226a88b55910e11ce4225e364b3e443f4f2fa98e