Fix SameSite WPT window.open() tests
These two tests had a typo in which the `origin` was used as the start
and the target of the test navigations. The `target` argument was unused.

This CL corrects that and updates the expected results for a couple
affected tests. (Plus some comments changes.)

Change-Id: I913f927c755ef21f2c6632bc84fca7b624a894a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3290927
Reviewed-by: Mike Taylor <miketaylr@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943722}
NOKEYCHECK=True
GitOrigin-RevId: 405e57fdbfa511283945745ea731b578d71f00ed