Window Placement: sort screens by x,y coordinates
The spec currently has an open issue about how to sort these.
Making an opinionated choice here to sort left to right, top to bottom.
After this patch, the spec will be updated to match.

Unfortunately, there is no easy way to test this.  The right place to
put this test is in window_placement_browsertest.cc.  These are the
only tests that can add and manipulate fake screens at runtime.
However, due to issue 1183791 these tests only run on ChromeOS.
Furthermore on ChromeOS, it turns out that the DisplayManager function
CreateDisplayFromDisplayInfoById resets the origin of any added
display and clobbers it with its own layout.  Therefore, there is no
platform that can currently set custom top/left of screens and have
them sort properly in a test.

In the future, when issue 1252062 is solved then we can add some WPT
internal tests that can do this work rather than trying to wrangle this
into a browser test.

For now, this was tested manually.

Bug: 897300, 1252062
Change-Id: I101a6eba9ab2590315deeb597567925877407a7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3293985
Commit-Queue: enne <enne@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943774}
NOKEYCHECK=True
GitOrigin-RevId: 61fdb6528272624ac232e9bb1ecee5a4860de9bd