Remove PerLineData array in TextDecorationInfo
We had an array for underline, overline and line-through lines
in TextDecorationInfo.
However as we always call SetPerLineData() and then ask for the bounds
or paint the line, we don't need to store that in different positions
on the array.

This patch removes the array, so we have a single PerLineData
in each TextDecorationInfo.
Apart from that this does a few other things:
* Renames PerLineData to LineData and SetPerLineData
  to SetLineData.
* Removes the "line" argument from a bunch of methods.
* Removes "line_" member from AppliedDecorationPainter
  as is no longer needed.

Change-Id: I6cb3a95bd9eb7f7e66ad99b9b48b8bae3a1d2681
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291658
Reviewed-by: Delan Azabani <dazabani@igalia.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Commit-Position: refs/heads/main@{#944083}
NOKEYCHECK=True
GitOrigin-RevId: 4bc308418011178e4b8f367dafae8aafd426b9c4