Properly populate top-level site for StorageKeys.
This makes sure the top-level site is properly populated in
NavigationRequest and RenderFrameHostImpl, by deriving the StorageKey
from the net::IsolationInfo associated with a frame and/or navigation
request.

This should have no behavioral changes unless the
ThirdPartyStoragePartitioning feature is turned on (which isn't by
default), as top-level site is not used anywhere without that feature
turned on.

We could consider not storing StorageKey separately in
RenderFrameHostImpl and always deriving it from the isolation info
when getting, but for this CL I kept the change as small as possible.

Also fixes the mojom traits to not assume that storage keys with a nonce
can't have a top-level site. Since those mojom traits were originally
written we changed StorageKey to populate top-level site even when a
nonce is present to be more consistent with IsolationInfo, so this
removes the no longer correct checks from the mojom traits.

Finally this also makes sure we don't lose the top-level site when
propagating it to blink.

As a follow-up we should clean up the various StorageKey and
BlinkStorageKey constructors and factory methods a bit; for example
CreateWithNonce is not used outside of tests at this point. Similar
some of the other constructors could probably be eliminated as well.

Bug: 1261329
Change-Id: I01f937332af398dc145ecb7f743a94b24ee9d7f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3232369
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944360}
NOKEYCHECK=True
GitOrigin-RevId: 74557ce8a270f6b8e21b9298f4121d4b910b9062