Ensure that <math> is properly exposed when MathMLCore is disabled
After [1], the MathML <math> element is no longer exposed with role
ROLE_SYSTEM_EQUATION on MSAA, causing regressions for Windows
assistive technologies on Chrome 96. This CL restores the original check
in AXNodeObject::NativeRoleIgnoringAria when MathMLCore is disabled. It
also extends test mathml/math.html so that the role of the <math>
element is tested on Windows APIs; as well as when MathMLCore is
disabled.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3168497

Bug: 6606, 1038895, 1051115, 1272563, 1272556
Change-Id: I79393b24ed24c9609af5cc6142cfae98fbaa86e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295404
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944404}
NOKEYCHECK=True
GitOrigin-RevId: b7d91d91fc1cfd152880fbd83a28aa4d0a6765a6