Remove incorrect CHECK from ImagePaintTimingDetector
We got some crashes about |pending_images_| not finding the record even
when the WeakPtr is not null. The likely cause of this is that erasing
from the HashMap does not necessarily immediately cause the destructor
of the unique_ptr from being called, and it's only then that the WeakPtr
becomes null.

Bug: 1271020
Change-Id: I6dca46d4ff7ff68df3564898e8c2ae95c570dab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3299176
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944701}
NOKEYCHECK=True
GitOrigin-RevId: a82f772230cc14e11616e48ac7e3727ba09c9372