EditContext: allow EditContext to connect to isolated elements
Per offline discussion during the TPAC week, the EditContext should be
allowed to connect to elements that are not in the tree, so it's easier
for the authors to prepare the tree. This CL changes the logic
accordingly (in Element::setEditContext, don't return if the element
is not in the tree, and in Element::RemovedFrom, don't reset
EditContext), and update the associated tests.

Note with this change, if an element has an associated EditContext, it
will be kept alive after removed from the tree. The authors are
responsible for the life time of both EditContext and the element.
Once the EditContext goes out of the scope, both EditContext and the
element will be GC'ed by Oilpan (as discussed in the comment
of CL#3180103 [1]

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3180103/comments/79c68113_2918cb3d

Bug: 999184
Change-Id: Iaa0a5d79b0e5f1d1a3df3568f1e45820173ef803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297169
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Alex Keng <shihken@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#944843}
NOKEYCHECK=True
GitOrigin-RevId: 77cb452cd341dca5642dc2cb8e18013d50e46b56