private-join-and-compute: Support component build for ChromeOS
private_membership library is depending on this library.
In order for it to support component build (in CL
crrev.com/c/3296024), private-join-and-compute library
need to support it as well.

This CL is adding the component build support for private
join and compute library only for ChromeOS platform.

Bug: b:207364544, b:207106194, b:201763017
Change-Id: I744566e7bf97e7dc4f1afb0d12eb7e4cab5c77c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3296027
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944983}
NOKEYCHECK=True
GitOrigin-RevId: 10fe65a5a7b9d2b7430a3afc601c53f543ff282f