Remove existing ADPF code
The production ADPF API is not accessible from sandboxed process
such as renderers. This means the existing code that was submitted
needs to be reworked.

Also the existing API in chromium does not allow for pipelining
which is important to ensure the timings are reported efficiently.

The clean rewrite has been uploaded as follow up CLs. This is
essentially reverting all previously submitted 3 CLs.

Bug: 1157620
Change-Id: Ibbeb693c62de8726997f9e8a7b08b63f1d950eec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295162
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945149}
NOKEYCHECK=True
GitOrigin-RevId: 3bd86e7be0b151c3d1a8793fc584dbf5bfbd4d70