Pre-paint: Only enter culled inlines in the relevant container fragments.
We used to enter them for pretty much every container fragment of the
inline formatting context, since we had no way of telling whether there
might be a descendant fragment item of the culled inline somewhere in
there or not.  While this is both wasteful [1] and messy, it was also
causing trouble for link highlighting, which needs to create
FragmentData entries even for culled inlines, if the highlighted link is
a culled inline. This CL is in preparation for fixing that. We also need
some fixes in the link highlighting code (which currently only uses the
legacy layout object tree). A follow-up CL will take care of this.

Just as for any other LayoutObject that we enter during pre-paint, we
need to know when we're entering it for the first and for the last time,
so that we initialize the FragmentData objects correctly (first time),
and clear paint flags (last time).

[1] There is a certain cost in checking whether a culled inline has
fragments for descendants inside a given container fragment, but that
should be cheaper than just always entering.

Bug: 829028
Change-Id: I4ae107121df204f5afbf28310338a5a65f8bd871
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3301039
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945303}
NOKEYCHECK=True
GitOrigin-RevId: 9cc23cb28d2d46f97a55ced69d853fef902c3983