Fix link highlighting for LayoutNGBlockFragmentation.
First of all, we need to make sure that during pre-paint, a non-atomic
inline gets a FragmentData object generated for each fragmentainer, even
if it's a culled inline.

Secondly, move away from legacy LayoutObject::OutlineRects(). This
method collects outlines from all fragmentainers in the stitched
"flow-thread" coordinate space. Slicing and clipping this into proper
columns afterwards is a tricky and unnecessary exercise, when we have
proper fragments for it.

Introducing FragmentDataIterator, which iterates over FragmentData
entries of a LayoutObject together with "corresponding" LayoutNG
structures, so that we know exactly what to process for each
fragmentainer, and where to paint it.

There's one remaining problem here, related to inline continuations. If
those get block-fragmented, they won't paint correctly, as we still fall
back to legacy OutlineRects() in such cases. This is hopefully
acceptable while we wait for LayoutNGBlockInInline to finish.

This fixes compositing/gestures/gesture-tapHighlight-multicol.html when
LayoutNGBlockFragmentation is enabled. Also added a couple of more tests
that would fail without this fix (with LayoutNGBlockFragmentation
enabled).

Bug: 1264798
Change-Id: I8e5278fd39a4ac9cec07907b65bd49a487d8e204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291443
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945320}
NOKEYCHECK=True
GitOrigin-RevId: 9d1f2e68eb2eae53f2743d8a0313b8c7aabcda74