[@container] Lift parenthesis handling to ConsumeInParens
In order to make it easier to define the grammar of container queries,
there was a change in the spec regarding where parens enclosing
a <media-feature> are handled. Previously, the parens were part
of <media-feature> itself, but as of [1] they were lifted to the
<media-in-parens> production.

This CL represents the corresponding change in our parser, as a
preparation for implementing the container queries syntax.
MediaQueryFeatureExpNode now represents the "bare" <media-feature>,
without any parens in its serialization. The parens are added
explicitly as a MediaQueryNestedExpNode wrapped around the
MediaQueryFeatureExpNode.

[1] https://github.com/w3c/csswg-drafts/pull/6806

Bug: 1214810
Change-Id: I69c0082fac02d18623beed06228be57a91e7c282
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3301699
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945339}
NOKEYCHECK=True
GitOrigin-RevId: 80c8bb264c8128885bf486e818129c4d2ecd5f47