Reduce flakiness of COOP tests by using timeouts
The COOP WPTs requiring a popup to be navigated were waiting before
navigating:
- that the onload event triggered
- requestAnimationFrame was executed, twice.

The latter is not always executed when the tab doesn't have the focus,
and gets executed as soon as it gets it. As an alternative, we added a
timeout, which avoids waiting for too long. It still achieves the same
goal in the tests I conducted than requestAnimationFrame: Make sure the
navigation is not considered as a redirection, and that a history entry
is created as a result.

Bug: 1233840
Change-Id: I5c8bbbccdc9b76ec32c795e7f5a4cdc594823e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297849
Auto-Submit: Pâris Meuleman <pmeuleman@chromium.org>
Commit-Queue: Pâris Meuleman <pmeuleman@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945351}
NOKEYCHECK=True
GitOrigin-RevId: 90387784185debba65db930b038aa42e697bf76c