[@container] Reject non-size features during parsing, fix orientation
Currently the ContainerQueryParser will accept any media feature.
This is not correct, it should instead only accept those explicitly
defined as size features [1]. This CL does this by plumbing a
FeatureSet object into the MediaQueryParser, which decides which
features are allowed.

Additionally, this CL makes the feature 'orientation' available for
use. It was valid to use it from the parser's perspective, but this
feature was not included in IsWidthDependent()/IsHeightDependent()
checks, which caused the QueriedAxes-check in ContainerQueryEvaluator::
Eval to fail.

Note that our current QueriedAxes-check in ContainerQueryEvaluator::
Eval is not quite according to what is currently specified. We
currently fail the query entirely if an uncontained axis is "mentioned"
by the query. Instead, individual features should evaluate to "unknown"
dynamically if the container does not have the right type. An
adjustment for this will come later.

[1] https://drafts.csswg.org/css-contain-3/#size-container

Fixed: 1269188
Change-Id: Icc08ece985ef22b377231a1073a6272aa967c6e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3301465
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#946044}
NOKEYCHECK=True
GitOrigin-RevId: f8bc29db2c139b45433b332bc1868fe978674dd1