Solve the case of the curiously absent DecoderStream trace events.
It turns out, gentle reader, that the trace events in DecoderStream
are not missing.  Rather, they never existed in the first place!

Via careful subterfuge, and manipulation of public records, the
devious code in RTCVideoDecoderStreamAdapter was able to pull the
wool over the Yard's eyes, reporting them to be missing when, in
fact, they were simply never there!  During the resulting confusion
of the Yard's efforts to track these supposed trace events, the
RTCVideoDecoderAdapter plotted its revenge against Mac OS hardware
video decoders, for reasons, as yet, unknown.

Rest assured, gentle reader, that we have unearthed this deviousness
and, with this CL add the trace events which were, previously,
never there.

As to the larger question of what malfeasance is intended towards
the law-abiding Mac RTC users, we shall continue our search.

Finally, there can be no argument, the game is afoot!

Bug: b/207764946
Change-Id: I27e7ac4f4d57d7809e3c4351316e3ced0182a1b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3305492
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#946066}
NOKEYCHECK=True
GitOrigin-RevId: b09acf5640cd9bbf9da0e7d0f68d3625063e1323