adpf: Hook up viz threads
This CL adds wrapper for accessing Android PerformanceHint (ADPF)
APIs. The APIs are accessed from the GPU process, and this CL adds
threads that are in the GPU process to the hint session. The threads
include the viz thread, IO thread, and GPU thread for display.

The target duration is hard coded to 12ms but is controllable by finch.
12ms is picked from the SurfaceFlinger deadline on pixel 6 devices.
The SF deadline is ~16ms for both 60 and 120hz modes, which should
include both CPU and GPU work. ADPF currently only works for CPU work
on threads that are added to the session, which means GPU work and
driver work that's not on the threads are not affected. So a fixed
portion is carved out of the deadline for GPU/driver work; this is the
same way HWUI models this work. So thus a default 12ms target duration
is selected.

The actual duration is the wall clock time from the vsync timemstamp to
swap_start. swap_start is close enough to the end of CPU work on the
display GPU thread. Queuing delay on these threads can be reduced by
ramping up CPU for these threads. And viz generally do not have
significant blocking synchronous calls during drawing which would
contribute to wall clock time but not be affected by ADPF, so there is
no need to account for them.

There is no security concern yet in this CL since it's all threads in
and reported by the GPU process.

Bug: 1157620
Change-Id: Ideae03cb9f6027bca6b9bb75ddb8e06755cc2459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295163
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#946549}
NOKEYCHECK=True
GitOrigin-RevId: cf15cf833a096442170b2d279a6c28caef08afcf