Introduce the DelayFirstParkingOfStrings trial.
The goal of the trial is to assess the performance impact of delaying
the first parking of strings by 60 seconds.

This delay was chosen by looking at UMA data and seeing that ~70%
of renderers do not make it past 1 min of lifetime.

The hope is that delaying the parking like this will not substantially
increase memory use but will save CPU cycles by avoiding the parking
of strings in renderers that are about to die.

Bug: 1229682
Change-Id: If425218afd942e1f8ebb85fc2d7985fbe1e97168
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3271249
Commit-Queue: Olivier Li <olivierli@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#946661}
NOKEYCHECK=True
GitOrigin-RevId: 4f69ad2960a2ab9e2d99b11f81b1956f84584697