[Paprika] Add onTouch event and plumb to VK API.
The new onTouch event will allow the virtual keyboard to react to touches
on a text field, which are currently not sent to the VK extension.
Main benefit is allowing for a handwriting layout after tapping with a pen.

AX-relnotes: nothing user visible

Bug: b:203019615
Change-Id: I961d90308aec4d4f2e8ebd913892f5dacfe3d861
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3249374
Commit-Queue: John Palmer <jopalmer@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Bao-Duy Tran <tranbaoduy@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#946844}
NOKEYCHECK=True
GitOrigin-RevId: ce7b60294d52ce6eb314d28f2705bbf59c08eb28